improvement: add pace penalty reset
chore: rename files and folders and restructure
