+++ VERSION.md
@@ -2,12 +2,13 @@
 
 | Version   | Changes |
 | :--- | :--- |
+| **Version 1.2.0** | **IMPROVEMENTS**<br>- **Reset Pace Penalty:** Added a new `Reset Pace Penalty` button to the portrait panel (next to the `Set Defenses` button). This lets you to manually reset your "high tide" movement mark mid-turn, if you have ended one action that is not being tracked by the module and are beginning another.<br><br>**UNDER THE HOOD**<br>- Version bump to confirm compatibility with FoundryVTT v14 with no changes to code needed.<br>- Renamed files and directories to match my other modules.|
 | **Version 1.1.4** | - Version bump to confirm compliance with Foundry's AI Content Policy.|
 | **Version 1.1.3** | - Added support and proper sorting for the new "Innate" spell list category.|
 | **Version 1.1.2** | - Fix to prevent Argon Combat HUD (CORE) giving a warning if a "Loot" actor is selected.|
 | **Version 1.1.1** | - Fix search bar filters in the Spell search bar. They should now show available instant and subconscious spells when toggled.|
 | **Version 1.1.0** | - Removed all spell attacks from the Attack panel.<br>- In the Spell panel (renamed from Spell (SCR) panel), if a spell has an attack, it shows both the SCR bonus and the OB bonus.<br>- If you cast a spell with an attack, it will change the state of the button after the attack to show it is armed and ready to be used again to make the attack roll.<br>- Clicking the button again (with your target selected / template placed) will do the attack and return the button to its unarmed state.<br>- The armed state has a slightly warmer background, a different icon, and only shows OB bonus (and the Place Template reminder for area spells).<br>- If the SCR roll fails (which the HUD cannot detect), you can dismiss the armed state by right-clicking the button.|
-| **Version 1.0.1** | - Moved to ApplicationV2.<br>- Added a check inside the 'Set Defenses' dialogue to warn users if they are trying to set Passive Dodge and Passive Block at the same time.<br>- Made sure that spell casting tooltips always appear above their buttons.|
+| **Version 1.0.1** | - Moved to ApplicationV2.<br>- Added a check inside the `Set Defenses` dialogue to warn users if they are trying to set Passive Dodge and Passive Block at the same time.<br>- Made sure that spell casting tooltips always appear above their buttons.|
 | **Version 1.0.0** | **PUBLIC RELEASE**<br>- Updated documentation.<br>- Updated master css theme.|
 | **Version 0.6.2** | - Added a reset to the pace penalty if an attack, spell or perception action are taken during the phase (this will only work if they are triggered by using the HUD, not from the character sheet). Pace penalty tracking then begins again in the next phase from 0.<br>- Fixed bugs in calculating AP cost for in-phase dedicated movement.<br>- All movement is now limited by encumbrance.<br>- The info box shows a warning if the token is spending more than 1 AP per phase (which could be true, but acts as a fail safe).<br>- Improved documentation of the movement HUD code to show calculation logic.|
 | **Version 0.6.1** | - Fixed bug in calculating AP cost for in-phase dedicated movement.<br>- Simplified movement HUD to a single track and a more detailed information box to show how far you can move until you incur another AP cost or move to a higher penalty.|
