+++ module.json
@@ -2,47 +2,19 @@
     "title": "Argon - Combat HUD (RMU)",
     "description": "Argon Combat HUD integration for the Rolemaster Unified (RMU) system",
     "id": "enhancedcombathud-rmu",
-    "version": "1.1.4",
+    "version": "1.2.0",
     "authors": [{ "name": "Ken Dearman" }],
-    "languages": [
-        {
-            "lang": "en",
-            "name": "English",
-            "path": "languages/en.json"
-        }
-    ],
-    "scripts": [],
-    "esmodules": ["index.js"],
+    "esmodules": ["main.js"],
     "styles": ["styles/filroden-theme.css", "styles/module.css"],
-    "packs": [],
     "relationships": {
-        "requires": [
-            {
-                "id": "enhancedcombathud",
-                "type": "module",
-                "compatibility": {
-                    "minimum": "4.1.0"
-                }
-            }
-        ],
-        "systems": [
-            {
-                "id": "rmu",
-                "type": "system",
-                "compatibility": {
-                    "minimum": "1.0.0"
-                }
-            }
-        ]
+        "requires": [{ "id": "enhancedcombathud", "type": "module", "compatibility": { "minimum": "4.1.0" } }],
+        "systems": [{ "id": "rmu", "type": "system", "compatibility": { "minimum": "1.0.0" } }]
     },
-    "compatibility": {
-        "minimum": "13",
-        "verified": "13.351"
-    },
+    "compatibility": { "minimum": "13", "verified": "14.359" },
     "url": "https://github.com/Filroden/enhancedcombathud-rmu",
     "bugs": "https://github.com/Filroden/enhancedcombathud-rmu/issues",
     "readme": "https://github.com/Filroden/enhancedcombathud-rmu/blob/main/README.md",
     "license": "https://mit-license.org/",
     "manifest": "https://github.com/Filroden/enhancedcombathud-rmu/releases/latest/download/module.json",
-    "download": " https://github.com/Filroden/enhancedcombathud-rmu/releases/download/v1.1.4/enhancedcombathud-rmu.zip"
+    "download": " https://github.com/Filroden/enhancedcombathud-rmu/releases/download/v1.2.0/enhancedcombathud-rmu.zip"
 }
